refactor(local): move to cli package (#2401)
* chore: move src & tests from local to cli

* chore: finish moving files

* refactor: finish moving/deleting files. Transpile and generate manifests before /cli unit tests to avoid oclif from doing it on every call of test.command.

* fix: version unit test due to file structure change and oclif load order.

* fix: drop node 14 since oclif does not support it and we are using their commands in unit tests now.

* chore: revert unneeded change

* chore: remove script that is no longer needed

* fix: local:version test approach
