+++ packages/cli/package.json
@@ -21,7 +21,6 @@
     "@heroku-cli/plugin-config": "^8.1.7",
     "@heroku-cli/plugin-container-registry-v5": "^8.1.7",
     "@heroku-cli/plugin-git": "^8.1.8",
-    "@heroku-cli/plugin-local": "^8.1.9",
     "@heroku-cli/plugin-oauth-v5": "^8.1.9",
     "@heroku-cli/plugin-orgs-v5": "^8.1.4",
     "@heroku-cli/plugin-pg-v5": "^8.1.7",
@@ -45,11 +44,11 @@
     "@oclif/plugin-which": "2.2.8",
     "debug": "4.1.1",
     "execa": "5.1.1",
+    "foreman": "^3.0.1",
     "fs-extra": "7.0.1",
     "http-call": "5.3.0",
     "netrc-parser": "3.1.6",
     "semver": "5.6.0",
-    "tslib": "1.14.1",
     "uuid": "3.3.2"
   },
   "devDependencies": {
@@ -65,6 +64,7 @@
     "@types/supports-color": "^5.3.0",
     "@types/write-json-file": "^3.2.1",
     "aws-sdk": "^2.421.0",
+    "bats": "^1.1.0",
     "chai": "^4.2.0",
     "globby": "^10.0.2",
     "lodash": "^4.17.11",
@@ -75,6 +75,7 @@
     "qqjs": "0.3.11",
     "read-pkg": "^4.0.1",
     "sinon": "^7.2.4",
+    "tslib": "1.14.1",
     "ts-node": "^10.9.1",
     "typescript": "4.8.4"
   },
@@ -119,7 +120,6 @@
       "@heroku-cli/plugin-config",
       "@heroku-cli/plugin-container-registry-v5",
       "@heroku-cli/plugin-git",
-      "@heroku-cli/plugin-local",
       "@heroku-cli/plugin-oauth-v5",
       "@heroku-cli/plugin-orgs-v5",
       "@heroku-cli/plugin-pg-v5",
@@ -279,9 +279,9 @@
     "build": "rm -rf lib && tsc",
     "postpublish": "rm -f oclif.manifest.json",
     "prepack": "yarn run build && oclif manifest",
-    "pretest": "tsc -p test --noEmit",
+    "pretest": "lerna run prepack --concurrency 4 && tsc -p test --noEmit",
     "test": "nyc mocha --forbid-only \"test/**/*.unit.test.ts\"",
-    "test:acceptance": "mocha --forbid-only \"test/**/*.acceptance.test.ts\"",
+    "test:acceptance": "mocha --forbid-only \"test/**/*.acceptance.test.ts\" && node ./bin/bats-test-runner",
     "test:smoke": "mocha --forbid-only \"test/**/smoke.acceptance.test.ts\"",
     "posttest": "yarn lint",
     "version": "oclif readme --multi && git add README.md ../../docs"
