+++ packages/cli/test/acceptance/plugin.acceptance.test.ts
@@ -1,4 +1,4 @@
-import execa from 'execa'
+import * as execa from 'execa'
 import * as fs from 'fs-extra'
 import * as path from 'path'
