--- packages/local/.gitignore
@@ -1,8 +1,0 @@
-*-debug.log
-*-error.log
-/.nyc_output
-/dist
-/lib
-/package-lock.json
-/tmp
-node_modules
