--- packages/local/.snyk
@@ -1,15 +1,0 @@
-# Snyk (https://snyk.io) policy file, patches or ignores known vulnerabilities.
-version: v1.25.0
-# ignores vulnerabilities until expiry date; change duration by modifying expiry date
-ignore:
-  SNYK-JS-ANSIREGEX-1583908:
-    - '*':
-        reason: None Given
-        expires: 2023-04-30T00:00:00.000Z
-        created: 2023-01-20T22:08:51.746Z
-  SNYK-JS-LODASHTEMPLATE-1088054:
-    - '*':
-        reason: None Given
-        expires: 2023-04-30T00:00:00.000Z
-        created: 2023-01-20T22:09:00.286Z
-patch: {}
