--- packages/local/CHANGELOG.md
@@ -1,259 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-## [8.1.9](https://github.com/heroku/cli/compare/v8.1.8...v8.1.9) (2023-06-21)
-
-**Note:** Version bump only for package @heroku-cli/plugin-local
-
-
-
-
-
-## [8.1.8](https://github.com/heroku/cli/compare/v8.1.7...v8.1.8) (2023-06-14)
-
-**Note:** Version bump only for package @heroku-cli/plugin-local
-
-
-
-
-
-## [8.1.4](https://github.com/heroku/cli/compare/v8.1.3...v8.1.4) (2023-05-24)
-
-**Note:** Version bump only for package @heroku-cli/plugin-local
-
-
-
-
-
-# [8.1.0](https://github.com/heroku/cli/compare/v8.0.6...v8.1.0) (2023-04-27)
-
-**Note:** Version bump only for package @heroku-cli/plugin-local
-
-
-
-
-
-## [8.0.6](https://github.com/heroku/cli/compare/v8.0.5...v8.0.6) (2023-04-24)
-
-**Note:** Version bump only for package @heroku-cli/plugin-local
-
-
-
-
-
-## [8.0.5](https://github.com/heroku/cli/compare/v8.0.4...v8.0.5) (2023-04-18)
-
-
-### Bug Fixes
-
-* oclif promote step ([#2301](https://github.com/heroku/cli/issues/2301)) ([4e94fb1](https://github.com/heroku/cli/commit/4e94fb1fdc1a591bd1b744e2da69c0e0df03ed6e))
-
-
-
-
-
-## [8.0.2](https://github.com/heroku/cli/compare/v7.69.1...v8.0.2) (2023-03-16)
-
-**Note:** Version bump only for package @heroku-cli/plugin-local
-
-
-
-
-
-## [7.69.2](https://github.com/heroku/cli/compare/v7.69.1...v7.69.2) (2023-03-16)
-
-**Note:** Version bump only for package @heroku-cli/plugin-local
-
-
-
-
-
-# [7.69.0](https://github.com/heroku/cli/compare/v7.68.3...v7.69.0) (2023-03-07)
-
-
-### Features
-
-* **local:** change default starting port ([#2261](https://github.com/heroku/cli/issues/2261)) ([88ff4d7](https://github.com/heroku/cli/commit/88ff4d74be7da5a8683869e9b612dea9dd069dfa))
-
-
-
-
-
-# [7.68.0](https://github.com/heroku/cli/compare/v8.0.1...v7.68.0) (2023-02-06)
-
-
-### Bug Fixes
-
-* revert to v7.67.2 ([#2235](https://github.com/heroku/cli/issues/2235)) ([0955a24](https://github.com/heroku/cli/commit/0955a24d6aeafdec7211ffd6179f772560f35098)), closes [#2231](https://github.com/heroku/cli/issues/2231) [#2230](https://github.com/heroku/cli/issues/2230) [#2229](https://github.com/heroku/cli/issues/2229) [#2228](https://github.com/heroku/cli/issues/2228) [#2227](https://github.com/heroku/cli/issues/2227) [#2225](https://github.com/heroku/cli/issues/2225) [#2144](https://github.com/heroku/cli/issues/2144) [#2216](https://github.com/heroku/cli/issues/2216) [#2207](https://github.com/heroku/cli/issues/2207) [#2212](https://github.com/heroku/cli/issues/2212) [#2212](https://github.com/heroku/cli/issues/2212)
-
-
-
-
-
-## [7.66.3](https://github.com/heroku/cli/compare/v7.66.2...v7.66.3) (2022-11-14)
-
-
-### Bug Fixes
-
-* debian builds to run in ci round 2 ([#2132](https://github.com/heroku/cli/issues/2132)) ([bd32cdc](https://github.com/heroku/cli/commit/bd32cdcee3a7a214b6aea6f309aae1a8ac2ae65e))
-
-
-
-
-
-# [7.60.0](https://github.com/heroku/cli/compare/v7.59.1...v7.60.0) (2022-03-23)
-
-**Note:** Version bump only for package @heroku-cli/plugin-local
-
-
-
-
-
-# [7.54.0](https://github.com/heroku/cli/compare/v7.47.10...v7.54.0) (2021-05-18)
-
-**Note:** Version bump only for package @heroku-cli/plugin-local
-
-
-
-
-
-# [7.53.0](https://github.com/heroku/cli/compare/v7.52.0...v7.53.0) (2021-04-27)
-
-**Note:** Version bump only for package @heroku-cli/plugin-local
-
-
-
-
-
-## [7.47.13](https://github.com/heroku/cli/compare/v7.47.12...v7.47.13) (2021-02-18)
-
-**Note:** Version bump only for package @heroku-cli/plugin-local
-
-
-
-
-
-# [7.43.0](https://github.com/heroku/cli/compare/v7.42.13...v7.43.0) (2020-09-15)
-
-**Note:** Version bump only for package @heroku-cli/plugin-local
-
-
-
-
-
-## [7.42.8](https://github.com/heroku/cli/compare/v7.42.7...v7.42.8) (2020-08-17)
-
-**Note:** Version bump only for package @heroku-cli/plugin-local
-
-
-
-
-
-# [7.42.0](https://github.com/heroku/cli/compare/v7.41.1...v7.42.0) (2020-06-03)
-
-**Note:** Version bump only for package @heroku-cli/plugin-local
-
-
-
-
-
-# [7.41.0](https://github.com/heroku/cli/compare/v7.40.0...v7.41.0) (2020-05-11)
-
-
-### Bug Fixes
-
-* **run:** fix run status code and use https ([#1351](https://github.com/heroku/cli/issues/1351)) ([ca8127f](https://github.com/heroku/cli/commit/ca8127fd5811d9f9c87aea891d33b36a6835a2ec))
-
-
-
-
-
-## [7.39.2](https://github.com/heroku/cli/compare/v7.39.1...v7.39.2) (2020-03-30)
-
-**Note:** Version bump only for package @heroku-cli/plugin-local
-
-
-
-
-
-## [7.38.1](https://github.com/heroku/cli/compare/v7.38.0...v7.38.1) (2020-02-10)
-
-**Note:** Version bump only for package @heroku-cli/plugin-local
-
-
-
-
-
-# [7.38.0](https://github.com/heroku/cli/compare/v7.37.0...v7.38.0) (2020-02-06)
-
-**Note:** Version bump only for package @heroku-cli/plugin-local
-
-
-
-
-
-# [7.36.0](https://github.com/heroku/cli/compare/v7.35.1...v7.36.0) (2020-01-20)
-
-**Note:** Version bump only for package @heroku-cli/plugin-local
-
-
-
-
-
-# [7.34.0](https://github.com/heroku/cli/compare/v7.33.3...v7.34.0) (2019-11-05)
-
-
-### Bug Fixes
-
-* **cli:** pin to typescript 3.3.3333 ([#1369](https://github.com/heroku/cli/issues/1369)) ([0488680](https://github.com/heroku/cli/commit/0488680))
-
-
-
-
-
-<a name="7.28.0"></a>
-# [7.28.0](https://github.com/heroku/cli/compare/v7.27.1...v7.28.0) (2019-08-19)
-
-
-### Bug Fixes
-
-* **local:** remove composite flag from tsconfig.json ([#1263](https://github.com/heroku/cli/issues/1263)) ([be9cdfb](https://github.com/heroku/cli/commit/be9cdfb))
-
-
-
-
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-## [7.26.2](https://github.com/heroku/cli/compare/v7.26.1...v7.26.2) (2019-06-27)
-
-
-### Bug Fixes
-
-* **local:** js files not being included in builds ([c60b2ce](https://github.com/heroku/cli/commit/c60b2ce))
-
-
-
-
-
-## [7.26.1](https://github.com/heroku/cli/compare/v7.26.0...v7.26.1) (2019-06-27)
-
-
-### Bug Fixes
-
-* **local:** fix foreman-procfile path ([32d1eae](https://github.com/heroku/cli/commit/32d1eae))
-
-
-
-
-
-# [7.26.0](https://github.com/heroku/cli/compare/v7.25.0...v7.26.0) (2019-06-26)
-
-
-### Features
-
-* **local:** add oclif version of local plugin ([#1252](https://github.com/heroku/cli/issues/1252)) ([eb77c7c](https://github.com/heroku/cli/commit/eb77c7c))
