--- packages/local/test/tsconfig.json
@@ -1,6 +1,0 @@
-{
-  "extends": "../tsconfig",
-  "compilerOptions": {
-    "noEmit": true
-  }
-}
